+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Changelog
 
 - Added SMASH support ([PR 2143](https://github.com/input-output-hk/daedalus/pull/2143))
 
+### Fixes
+
+- Disabled ITN rewards redemption in case Daedalus is not fully in sync ([PR 2146](https://github.com/input-output-hk/daedalus/pull/2146))
+
 ## 2.1.0
 
 ### Features